Remove rust dependency
Interoperating with rust made the build scripts complicated and
prevented this project from building with ocaml 5. This change replaces
the cpal-based audio player with a libao-based player, for which there
exists ocaml bindings.

A side effect of this change is that the project loses support for live
midi playback, wav decoding, and the ability to run on windows. This
support will be gradually added back in future changes without
reintroducing the rust dependency.